+++ src/test/java/pt/flow/decision/FlowWorkbench.java
@@ -1,8 +1,8 @@
-package pt.marcioneves.decision;
+package pt.flow.decision;
 
 import org.junit.Test;
-import pt.marcioneves.decision.core.AbstractContext;
-import pt.marcioneves.decision.core.ICommand;
+import pt.flow.decision.core.AbstractContext;
+import pt.flow.decision.core.ICommand;
 
 import java.util.Random;
 import java.util.logging.Logger;
